+++ src/main/java/illyena/gilding/avengers/client/render/blockentity/TeleportAnchorBlockEntityRenderer.java
@@ -26,7 +26,7 @@ public class TeleportAnchorBlockEntityRenderer implements BlockEntityRenderer<Te
     public void render(TeleportAnchorBlockEntity entity, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, int overlay) {
         if (entity.isRecentlyGenerated()) {
             float g = entity.getRecentlyGeneratedBeamHeight(tickDelta);
-            double d = (double) entity.getWorld().getTopY();
+            double d = entity.getWorld().getTopY();
             g = MathHelper.sin(g * 3.1415927F);
             int k = MathHelper.floor((double) g * d);
             long l = entity.getWorld().getTime();
